Refactor project structure and update configurations
- Adjusted createPages.js to change directory paths for page imports and creation.
- Updated next.config.js to set the app directory to 'src' and specified a custom distribution directory.
- Modified package.json to include cross-env for environment variable management in scripts.
- Enhanced tsconfig.json to include new type definitions for the 'dist' directory.
- Removed unused files and directories from the app structure, including layout, favicon, and page components.
- Updated yarn.lock to reflect new dependencies and removed obsolete entries.
