+++ package.json
@@ -14,9 +14,9 @@
     "codegen": "cd src/contracts/ && ${AZTEC_BUILDER:-aztec-builder} codegen ./target -o ../../artifacts",
     "clean": "rm -rf ./dist .tsbuildinfo ./artifacts ./src/contracts/target",
     "prep": "yarn clean && yarn compile && yarn codegen",
-    "dev": "next dev",
-    "build": "yarn prep && next build",
-    "start": "next start",
+    "dev": "cross-env NEXT_PUBLIC_APP_DIR=src next dev",
+    "build": "yarn prep && cross-env NEXT_PUBLIC_APP_DIR=src next build",
+    "start": "cross-env NEXT_PUBLIC_APP_DIR=src next start",
     "formatting": "prettier --check ./src && eslint ./src",
     "formatting:fix": "prettier -w ./src",
     "test": "yarn test:node && yarn test:browser",
@@ -67,6 +67,7 @@
     "@typescript-eslint/parser": "^6.0.0",
     "assert": "^2.1.0",
     "autoprefixer": "^10.4.21",
+    "cross-env": "^7.0.3",
     "eslint": "^8.21.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-import-resolver-typescript": "^3.5.5",
