+++ src/app/bridge/page.tsx
@@ -1,5 +1,5 @@
 import AppLayout from '../components/AppLayout';
-import BridgeContent from '../../src/pages/Bridge';
+import BridgeContent from '../../../src/pages/Bridge';
 
 export default function BridgePage() {
   return (
