+++ src/app/buy/page.tsx
@@ -1,5 +1,5 @@
 import AppLayout from '../components/AppLayout';
-import BuyContent from '../../src/pages/Buy';
+import BuyContent from '../../../src/pages/Buy';
 
 export default function BuyPage() {
   return (
