+++ src/app/favicon/favicon.ico
