+++ src/app/listing/page.tsx
@@ -1,5 +1,5 @@
 import AppLayout from '../components/AppLayout';
-import ListingContent from '../../src/pages/Listing';
+import ListingContent from '../../../src/pages/Listing';
 
 export default function ListingPage() {
   return (
