+++ src/app/mint/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import AppLayout from '../components/AppLayout';
-import MintContent from '../../src/pages/Mint';
+import MintContent from '../../../src/pages/Mint';
 
 export default function MintPage() {
     return (
