+++ src/app/transfer/page.tsx
@@ -1,5 +1,5 @@
 import AppLayout from '../components/AppLayout';
-import TransferContent from '../../src/pages/Transfer';
+import TransferContent from '../../../src/pages/Transfer';
 
 export default function TransferPage() {
   return (
