+++ src/app/verify-ownership/page.tsx
@@ -1,5 +1,5 @@
 import AppLayout from '../components/AppLayout';
-import VerifyOwnershipContent from '../../src/pages/VerifyOwnership';
+import VerifyOwnershipContent from '../../../src/pages/VerifyOwnership';
 
 export default function VerifyOwnershipPage() {
   return (
